fix: remove sleep-rfkill service
Although this works on most devices, it unfortunately does not
always work on the SuiPlay0X1. Sometimes the device will try to
enter S2 sleep, the sleep-rfkill service immediately fails due to
a hardware issue, and then the device falls back to S0 sleep which
uses a lot of power since the device remains on.