feat(plymouth): remove 'playtron' logo
It does not look great and we cannot customize it. Grid already
displays a 'playtron' logo animation when it starts.

The most reliable way I could get this to work was to replace the
watermark with a 1x1 black pixel image. I have only seen black BIOS
screens so this should not be an issue. We also get to keep the
spinner to let the user know that something is happening.